Separate code into UI, TaskList and Storage
